requirements.txt
pywin32
pywin32-ctypes
pygame
backoff
loguru
json2txttree
python-dateutil
humanize
psutil